Add files via upload
took out all the pre-existing ducks and hard coded in 5 ducks. There's code for the fish, but they're a bit wackily placed :/ They are not in the scene for now. I also have code to randomly have the fish appear but I think the lag makes it so that they all appear at the same time instead of one by one